用户端菜品浏览相关代码开发

+++ sky-server/src/main/java/com/sky/controller/admin/CategoryController.java
@@ -99,7 +99,7 @@ public class CategoryController {
      */
     @GetMapping("/list")
     @ApiOperation("查询分类")
-    public Result<List<Category>> list(Long type){
+    public Result<List<Category>> list(Integer type){
         log.info("查询类型数量：{} ",type);
         List<Category> list = categoryService.list(type);
         return Result.success(list);
